Quick refactor of the utils section for tidiness
