+++ src/utils/internal-data/dataLastUpdated.ts
@@ -1,7 +1,7 @@
 
 /** The date when the data for the given year was last updated - will clear local caching */
 export const dataLastUpdated: Record<string, number> = { //Update by executing ./get-epoch-time.sh
-  "Men_2015/6": 1572354347,
-  "Men_2018/9": 1572354347,
-  "Women_2018/9": 1572354348
+  "Men_2015/6": 1572354349,
+  "Men_2018/9": 1572354349,
+  "Women_2018/9": 1572354349
 };
