+++ src/utils/public-data/efficiencyAverages.ts
